+++ src/spec.ts
@@ -1,14 +1,6 @@
-import {ValidationError} from "./validation_error";
+import {ValidationError, ValidationFailure} from "./validation_error";
+export {ValidationFailure} from "./validation_error";
 
-
-export interface ValidationFailure {
-    code: string;
-    value: unknown;
-    allowed?: unknown;
-    message: string;
-    key?: string | number;
-    nestedErrors?: ValidationFailure[];
-}
 
 export interface GlobalOptions {
     readonly failEarly?: boolean;